// Neptune.Web/src/app/shared/generated/model/maintenance-record-detail-dto.ts
export class MaintenanceRecordDetailDto { 
    MaintenanceRecordID?: number;
    FieldVisitID?: number;
    TreatmentBMPID?: number;
    TreatmentBMPTypeID?: number;
    MaintenanceRecordTypeID?: number | null;
    MaintenanceRecordTypeDisplayName?: string | null;
    MaintenanceRecordDescription?: string | null;
    Observations?: Array<MaintenanceRecordObservationDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}