// Neptune.Web/src/app/shared/generated/model/maintenance-record-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MaintenanceRecordObservationUpsertDto } from './maintenance-record-observation-upsert-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class MaintenanceRecordUpsertDto { 
    MaintenanceRecordTypeID?: number | null;
    MaintenanceRecordDescription?: string | null;
    Observations?: Array<MaintenanceRecordObservationUpsertDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface MaintenanceRecordUpsertDtoForm { 
    MaintenanceRecordTypeID?: FormControl<number>;
    MaintenanceRecordDescription?: FormControl<string>;
    Observations?: FormControl<Array<MaintenanceRecordObservationUpsertDto>>;
}

export class MaintenanceRecordUpsertDtoFormControls { 
    public static MaintenanceRecordTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static MaintenanceRecordDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Observations = (value: FormControlState<Array<MaintenanceRecordObservationUpsertDto>> | Array<MaintenanceRecordObservationUpsertDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<MaintenanceRecordObservationUpsertDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}