// Neptune.Web/src/app/shared/generated/model/maintenance-record-observation-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class MaintenanceRecordObservationUpsertDto { 
    CustomAttributeTypeID?: number;
    Values?: Array<string> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface MaintenanceRecordObservationUpsertDtoForm { 
    CustomAttributeTypeID?: FormControl<number>;
    Values?: FormControl<Array<string>>;
}

export class MaintenanceRecordObservationUpsertDtoFormControls { 
    public static CustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Values = (value: FormControlState<Array<string>> | Array<string> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<string>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}